requirements.txt
dlib
numpy
opencv-python
pandas
Pillow
scikit-image
scipy